moved AnswerFilter into conversations package
